+++ domain/actionSpace/gatherWood.unit.ts
@@ -4,7 +4,7 @@ import { Dwarf, Player } from "../entity/Player";
 import { ActionSpace, ActionSpaceId } from "../entity/ActionSpace";
 import { Resources, ResourceType } from "../entity/Resources";
 import { Game } from "../entity/Game";
-import { buildBaseObjects, expectMutationsOfType, shouldPlaceDwarf } from "../testUtils";
+import { buildBaseObjects, expectMutationsOfType, shouldPlaceDwarf } from "../util/tests";
 
 describe("Gather wood", () => {
     shouldPlaceDwarf(GatherWood.execute);
